test db connection 14

+++ api.py
@@ -19,7 +19,6 @@ mydb = mysql.connector.connect(
 mycursor = mydb.cursor()
 
 def studentExists(uid):
-    '''
     sql = "SELECT * FROM students WHERE uid = %s"
     val = (uid,)
     mycursor.execute(sql, val)
@@ -28,7 +27,7 @@ def studentExists(uid):
         return True
     else:
         return False
-    '''
+    
 
 # Serves GET and POST requests for API
 @app.route("/", methods=['GET'])
@@ -37,8 +36,6 @@ def getHomePage():
 
 @app.route("/students", methods=['GET'])
 def getAllStudents():
-    return "<h1>get all students</h1>", 200
-    '''
     sql = "SELECT * FROM students"
     mycursor.execute(sql)
 
@@ -49,12 +46,11 @@ def getAllStudents():
     for result in myresult:
         json_data.append(dict(zip(row_headers,result)))
         
-    return jsonify(json_data), 200 '''
+    return jsonify(json_data), 200 
 
 @app.route("/student/<uid>", methods=['GET'])
 def getStudent(uid):
-    return "<h1>GET STUDENT</h1>", 200
-    '''if not studentExists(uid):
+    if not studentExists(uid):
         return {}, 404
 
     sql = "SELECT * FROM students WHERE uid = %s"
@@ -65,13 +61,11 @@ def getStudent(uid):
     json_data =[]
     for result in myresult:
         json_data.append(dict(zip(row_headers,result)))
-    return jsonify(json_data[0]), 200'''
+    return jsonify(json_data[0]), 200
 
 @app.route("/add", methods=['POST'])
 def addStudent():
-    return "<h1>ADD STUDENT</h1>", 200
-
-    '''student = request.json
+    student = request.json
     uid = student['uid']
     name = student['name']
     gender = student['gender']
@@ -84,20 +78,18 @@ def addStudent():
         mydb.commit()
         return jsonify({'success':True}), 200
     else:
-        return {}, 400'''
+        return {}, 400
 
 @app.route("/delete/<uid>", methods=['DELETE'])
 def deleteStudent(uid):
-    return "<h1>DELETE STUDENT</h1>", 200
-
-    '''if  studentExists(uid):
+    if  studentExists(uid):
         sql = "DELETE FROM students WHERE uid = %s"
         val = (uid, )
         mycursor.execute(sql, val)        
         mydb.commit()
         return jsonify({'success':True}), 200
     else:
-        return {} , 404'''
+        return {} , 404
         
 
 # Running application
